+++ fastcrypto-cli/Cargo.toml
@@ -19,8 +19,6 @@ exitcode = "1.1.2"
 [dev-dependencies]
 assert_cmd = "2.0.0"
 regex = "1.5.4"
-escargot = "0.5.7"
-lazy_static = "1.4.0"
 
 [[bin]]
 name = "ecvrf-cli"
